requirements.txt
amqp==5.1.1
asgiref==3.6.0
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
Django==4.1.6
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
Markdown==3.4.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.2
paramiko==3.0.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyNaCl==1.5.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
